refactor!: move Fu enum into separate module
This makes `lib.rs` smaller and easier to maintain.
